Use static methods
