+++ .eslintrc
@@ -18,6 +18,7 @@
     "import/no-extraneous-dependencies": "off",
     "import/no-internal-modules": "off",
     "import/no-named-as-default": "off",
+    "import/prefer-default-export": "off",
     "import/order": "error",
     "prettier/prettier": ["error"],
     "react/jsx-one-expression-per-line": "off",
