+++ src/jobs/Cart/CartQueuedJobs.ts
@@ -22,7 +22,7 @@ export class CartQueuedJobs extends QueuedJobsHandler<ErrorCartTypes> {
   }
 
   setCartItem = async () => {
-    const checkout = this.localStorageHandler.getCheckout();
+    const checkout = LocalStorageHandler.getCheckout();
 
     if (checkout) {
       const { data, error } = await this.apolloClientManager.setCartItem(
